move MODEL_METADATA to its own module
