+++ src/natcap/invest/coastal_blue_carbon/preprocessor.py
@@ -12,11 +12,10 @@ from .. import utils
 from .. import spec_utils
 from ..spec_utils import u
 from .. import validation
-from .. import MODEL_METADATA
-from .. import install_locale
+from ..model_metadata import MODEL_METADATA
+from .. import gettext
 from . import coastal_blue_carbon
 
-gettext = install_locale()
 LOGGER = logging.getLogger(__name__)
 
 ARGS_SPEC = {
