+++ src/natcap/invest/coastal_vulnerability.py
@@ -25,10 +25,10 @@ from . import utils
 from . import spec_utils
 from .spec_utils import u
 from . import validation
-from . import MODEL_METADATA
-from . import install_locale
+from .model_metadata import MODEL_METADATA
+from . import gettext
 
-gettext = install_locale()
+
 LOGGER = logging.getLogger(__name__)
 
 POLYLINE_VECTOR_MSG = gettext('Must be a polyline vector')
