+++ src/natcap/invest/delineateit/delineateit.py
@@ -18,11 +18,11 @@ from .. import utils
 from .. import spec_utils
 from ..spec_utils import u
 from .. import validation
-from .. import MODEL_METADATA
-from .. import install_locale
+from ..model_metadata import MODEL_METADATA
+from .. import gettext
 from . import delineateit_core
 
-gettext = install_locale()
+
 LOGGER = logging.getLogger(__name__)
 
 ARGS_SPEC = {
