+++ src/natcap/invest/forest_carbon_edge_effect.py
@@ -20,10 +20,10 @@ from . import utils
 from . import spec_utils
 from .spec_utils import u
 from . import validation
-from . import MODEL_METADATA
-from . import install_locale
+from .model_metadata import MODEL_METADATA
+from . import gettext
 
-gettext = install_locale()
+
 LOGGER = logging.getLogger(__name__)
 
 # grid cells are 100km. Becky says 500km is a good upper bound to search
