+++ src/natcap/invest/globio.py
@@ -11,14 +11,14 @@ from osgeo import gdal
 from osgeo import ogr
 from osgeo import osr
 
-from . import MODEL_METADATA
+from .model_metadata import MODEL_METADATA
 from . import spec_utils
 from . import utils
 from . import validation
 from .spec_utils import u
-from . import install_locale
+from . import gettext
 
-gettext = install_locale()
+
 LOGGER = logging.getLogger(__name__)
 
 # this value of sigma == 9.0 was derived by Justin Johnson as a good
