+++ src/natcap/invest/recreation/recmodel_client.py
@@ -34,10 +34,10 @@ from .. import utils
 from .. import spec_utils
 from ..spec_utils import u
 from .. import validation
-from .. import MODEL_METADATA
-from .. import install_locale
+from ..model_metadata import MODEL_METADATA
+from .. import gettext
 
-gettext = install_locale()
+
 LOGGER = logging.getLogger(__name__)
 
 # This URL is a NatCap global constant
