+++ src/natcap/invest/routedem.py
@@ -12,10 +12,10 @@ from . import utils
 from . import spec_utils
 from .spec_utils import u
 from . import validation
-from . import MODEL_METADATA
-from . import install_locale
+from .model_metadata import MODEL_METADATA
+from . import gettext
 
-gettext = install_locale()
+
 LOGGER = logging.getLogger(__name__)
 
 INVALID_BAND_INDEX_MSG = gettext('Must be between 1 and {maximum}')
