+++ src/natcap/invest/stormwater.py
@@ -12,10 +12,10 @@ from . import spec_utils
 from .spec_utils import u
 from . import utils
 from . import validation
-from . import MODEL_METADATA
-from . import install_locale
+from .model_metadata import MODEL_METADATA
+from . import gettext
 
-gettext = install_locale()
+
 LOGGER = logging.getLogger(__name__)
 
 # a constant nodata value to use for intermediates and outputs
