+++ src/natcap/invest/ui/annual_water_yield.py
@@ -1,7 +1,8 @@
 # coding=UTF-8
 
 from natcap.invest.ui import model, inputs
-from natcap.invest import annual_water_yield, MODEL_METADATA
+from natcap.invest.model_metadata import MODEL_METADATA
+from natcap.invest import annual_water_yield
 
 
 class AnnualWaterYield(model.InVESTModel):
