+++ src/natcap/invest/ui/coastal_vulnerability.py
@@ -1,7 +1,8 @@
 # coding=UTF-8
 
 from natcap.invest.ui import model, inputs
-from natcap.invest import coastal_vulnerability, MODEL_METADATA
+from natcap.invest.model_metadata import MODEL_METADATA
+from natcap.invest import coastal_vulnerability
 
 from osgeo import gdal
