+++ src/natcap/invest/ui/habitat_quality.py
@@ -1,7 +1,8 @@
 # coding=UTF-8
 
 from natcap.invest.ui import model, inputs
-from natcap.invest import habitat_quality, MODEL_METADATA
+from natcap.invest.model_metadata import MODEL_METADATA
+from natcap.invest import habitat_quality
 
 
 class HabitatQuality(model.InVESTModel):
