+++ src/natcap/invest/ui/hra.py
@@ -1,6 +1,7 @@
 # coding=UTF-8
 from natcap.invest.ui import model, inputs
-from natcap.invest import hra, MODEL_METADATA
+from natcap.invest.model_metadata import MODEL_METADATA
+from natcap.invest import hra
 
 
 class HabitatRiskAssessment(model.InVESTModel):
