+++ src/natcap/invest/ui/routedem.py
@@ -1,7 +1,8 @@
 # coding=UTF-8
 
 from natcap.invest.ui import model, inputs
-from natcap.invest import routedem, MODEL_METADATA
+from natcap.invest.model_metadata import MODEL_METADATA
+from natcap.invest import routedem
 
 
 class RouteDEM(model.InVESTModel):
