+++ src/natcap/invest/ui/urban_cooling_model.py
@@ -1,7 +1,8 @@
 # coding=UTF-8
 
 from natcap.invest.ui import model, inputs
-from natcap.invest import urban_cooling_model, MODEL_METADATA
+from natcap.invest.model_metadata import MODEL_METADATA
+from natcap.invest import urban_cooling_model
 
 
 class UrbanCoolingModel(model.InVESTModel):
