+++ src/natcap/invest/ui/urban_flood_risk_mitigation.py
@@ -1,7 +1,8 @@
 # coding=UTF-8
 
 from natcap.invest.ui import model, inputs
-from natcap.invest import urban_flood_risk_mitigation, MODEL_METADATA
+from natcap.invest.model_metadata import MODEL_METADATA
+from natcap.invest import urban_flood_risk_mitigation
 
 
 class UrbanFloodRiskMitigation(model.InVESTModel):
