+++ src/natcap/invest/urban_flood_risk_mitigation.py
@@ -17,10 +17,10 @@ from . import utils
 from . import spec_utils
 from .spec_utils import u
 from . import validation
-from . import MODEL_METADATA
-from . import install_locale
+from .model_metadata import MODEL_METADATA
+from . import gettext
 
-gettext = install_locale()
+
 LOGGER = logging.getLogger(__name__)
 
 ARGS_SPEC = {
