+++ tests/test_args_specs.py
@@ -2,7 +2,7 @@ import importlib
 import re
 import unittest
 
-from natcap.invest import MODEL_METADATA
+from natcap.invest.model_metadata import MODEL_METADATA
 import pint
